FinalProject/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
baidu-aip==4.16.11
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
Pillow==9.5.0
playsound==1.3.0
PyAudio==0.2.13
pydub==0.25.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
qt-material==2.14
requests==2.30.0
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
pyobjc
AppKit